data_scrapper/requirements.txt
huggingface_hub
requests
python-dotenv
opencv-python-headless
